refactor: is moving getMarkupForSvg into separate helper file
